cleanup: cleanup logs and remove duplicated code
