+++ backend/app/app.go
@@ -98,10 +98,6 @@ func NewApp(config internal.Configuration) (*App, error) {
 	gridClient, err := deployer.NewTFPluginClient(
 		config.SystemAccount.Mnemonic,
 		deployer.WithNetwork(config.SystemAccount.Network),
-		// TODO: remove this after testing
-		deployer.WithSubstrateURL("wss://tfchain.dev.grid.tf/ws"),
-		deployer.WithProxyURL("https://gridproxy.dev.grid.tf"),
-		deployer.WithRelayURL("wss://relay.dev.grid.tf"),
 	)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create TF grid client: %w", err)
@@ -238,10 +234,6 @@ func (app *App) registerHandlers() {
 				deploymentGroup.POST("/:name/nodes", app.handlers.HandleAddNode)
 				deploymentGroup.DELETE("/:name/nodes/:node_name", app.handlers.HandleRemoveNode)
 			}
-
-			// TODO: Task routes
-			// deployerGroup.GET("/tasks", app.handlers.ListUserTasksHandler)
-			// deployerGroup.GET("/tasks/:task_id", app.handlers.GetTaskStatusHandler)
 
 			// Notification routes
 			deployerGroup.GET("/notifications", app.handlers.GetNotificationsHandler)
