+++ backend/tests/deploy_test.go
@@ -23,7 +23,7 @@ func TestClient_DeployCluster(t *testing.T) {
 	t.Log("Login successful")
 
 	cluster := kubedeployer.Cluster{
-		Name:  "jrk8s02",
+		Name:  "mycluster",
 		Token: "test-token-123",
 		Nodes: []kubedeployer.Node{
 			{
@@ -36,7 +36,7 @@ func TestClient_DeployCluster(t *testing.T) {
 				EnvVars: map[string]string{
 					"SSH_KEY": "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQDJ1t4Ug8EfykmJwAbYudyYYN/f7dZaVg3KGD2Pz0bd9pajAAASWYrss3h2ctCZWluM6KAt289RMNzxlNUkOMJ9WhCIxqDAwtg05h/J27qlaGCPP8BCEITwqNKsLwzmMZY1UFc+sSUyjd35d3kjtv+rzo2meaReZnUFNPisvxGoygftAE6unqNa7TKonVDS1YXzbpT8XdtCV1Y6ACx+3a82mFR07zgmY4BVOixNBy2Lzpq9KiZTz91Bmjg8dy4xUyWLiTmnye51hEBgUzPprjffZByYSb2Ag9hpNE1AdCGCli/0TbEwFn9iEroh/xmtvZRpux+L0OmO93z5Sz+RLiYXKiYVV5R5XYP8y5eYi48RY2qr82sUl5+WnKhI8nhzayO9yjPEp3aTvR1FdDDj5ocB7qKi47R8FXIuwzZf+kJ7ZYmMSG7N21zDIJrz6JGy9KMi7nX1sqy7NSqX3juAasIjx0IJsE8zv9qokZ83hgcDmTJjnI+YXimelhcHn4M52hU= omar@jarvis",
 				},
-				NodeID: 337,
+				NodeID: 5133,
 			},
 			{
 				Name:     "master",
@@ -48,10 +48,10 @@ func TestClient_DeployCluster(t *testing.T) {
 				EnvVars: map[string]string{
 					"SSH_KEY": "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQDJ1t4Ug8EfykmJwAbYudyYYN/f7dZaVg3KGD2Pz0bd9pajAAASWYrss3h2ctCZWluM6KAt289RMNzxlNUkOMJ9WhCIxqDAwtg05h/J27qlaGCPP8BCEITwqNKsLwzmMZY1UFc+sSUyjd35d3kjtv+rzo2meaReZnUFNPisvxGoygftAE6unqNa7TKonVDS1YXzbpT8XdtCV1Y6ACx+3a82mFR07zgmY4BVOixNBy2Lzpq9KiZTz91Bmjg8dy4xUyWLiTmnye51hEBgUzPprjffZByYSb2Ag9hpNE1AdCGCli/0TbEwFn9iEroh/xmtvZRpux+L0OmO93z5Sz+RLiYXKiYVV5R5XYP8y5eYi48RY2qr82sUl5+WnKhI8nhzayO9yjPEp3aTvR1FdDDj5ocB7qKi47R8FXIuwzZf+kJ7ZYmMSG7N21zDIJrz6JGy9KMi7nX1sqy7NSqX3juAasIjx0IJsE8zv9qokZ83hgcDmTJjnI+YXimelhcHn4M52hU= omar@jarvis",
 				},
-				NodeID: 179,
+				NodeID: 5133,
 			},
 			{
-				Name:     "worker2",
+				Name:     "worker",
 				Type:     kubedeployer.NodeTypeWorker,
 				CPU:      1,
 				Memory:   2 * 1024,
@@ -93,10 +93,10 @@ func TestClient_AddNode(t *testing.T) {
 		EnvVars: map[string]string{
 			"SSH_KEY": "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQDJ1t4Ug8EfykmJwAbYudyYYN/f7dZaVg3KGD2Pz0bd9pajAAASWYrss3h2ctCZWluM6KAt289RMNzxlNUkOMJ9WhCIxqDAwtg05h/J27qlaGCPP8BCEITwqNKsLwzmMZY1UFc+sSUyjd35d3kjtv+rzo2meaReZnUFNPisvxGoygftAE6unqNa7TKonVDS1YXzbpT8XdtCV1Y6ACx+3a82mFR07zgmY4BVOixNBy2Lzpq9KiZTz91Bmjg8dy4xUyWLiTmnye51hEBgUzPprjffZByYSb2Ag9hpNE1AdCGCli/0TbEwFn9iEroh/xmtvZRpux+L0OmO93z5Sz+RLiYXKiYVV5R5XYP8y5eYi48RY2qr82sUl5+WnKhI8nhzayO9yjPEp3aTvR1FdDDj5ocB7qKi47R8FXIuwzZf+kJ7ZYmMSG7N21zDIJrz6JGy9KMi7nX1sqy7NSqX3juAasIjx0IJsE8zv9qokZ83hgcDmTJjnI+YXimelhcHn4M52hU= omar@jarvis",
 		},
-		NodeID: 150,
+		NodeID: 6244,
 	}
 
-	taskID, err := client.AddNode("jrk8s02", newNode)
+	taskID, err := client.AddNode("mycluster", newNode)
 	if err != nil {
 		t.Errorf("Add node failed: %v", err)
 		return
@@ -114,7 +114,7 @@ func TestClient_RemoveNode(t *testing.T) {
 	}
 	t.Log("Login successful")
 
-	err = client.RemoveNode("jrk8s02", "worker2")
+	err = client.RemoveNode("mycluster", "worker")
 	if err != nil {
 		t.Errorf("Remove node failed: %v", err)
 		return
@@ -130,7 +130,7 @@ func TestClient_DeleteCluster(t *testing.T) {
 		return
 	}
 
-	err := client.DeleteCluster("jrk8s02")
+	err := client.DeleteCluster("mycluster")
 	if err != nil {
 		t.Errorf("Failed to delete cluster: %v", err)
 		return
